Code improvement with rubocop

+++ spec/requests/api/members_spec.rb
@@ -4,36 +4,35 @@ require 'rails_helper'
 
 RSpec.describe 'Members', type: :request do
   let (:attributes) { attributes_for :member }
-  
+
   describe '.index' do
     context "when member's table is empty" do
       before { get '/api/members' }
 
-      it 'should return a HTTP STATUS 200' do
+      it 'does return a HTTP STATUS 200' do
         expect(response).to have_http_status(:ok)
       end
 
-      it 'should return an empty array' do
+      it 'does return an empty array' do
         expect(json_response[:members]).to eq([])
       end
     end
 
     context "when member's table is not empty" do
-
       before do
-        10.times do create(:member, attributes) end
-          get '/api/members'
+        10.times { create(:member, attributes) }
+        get '/api/members'
       end
 
-      it 'should return a HTTP STATUS 200' do
+      it 'does return a HTTP STATUS 200' do
         expect(response).to have_http_status(:ok)
       end
 
-      it 'should return an array of members' do
+      it 'does return an array of members' do
         expect(json_response[:members].length).to eq(10)
       end
 
-      it 'each member should have a name and description' do
+      it 'each member does have a name and description' do
         check_keys(json_response[:members])
       end
     end
@@ -51,20 +50,18 @@ RSpec.describe 'Members', type: :request do
       context 'when parameters are valid' do
 
         before do |example|
-          unless example.metadata[:skip_before]
-            create_member(attributes, @token)
-          end
+          create_member(attributes, @token) unless example.metadata[:skip_before]
         end
 
-        it 'should add a member to database', :skip_before do
+        it 'does add a member to database', :skip_before do
           expect{ create_member(attributes, @token) }.to change(Member, :count).by(1)
         end
 
-        it 'should return a HTTP STATUS 200' do
+        it 'does return a HTTP STATUS 200' do
           expect(response).to have_http_status(:created)
         end
 
-        it 'should return the created member' do
+        it 'does return the created member' do
           member = Member.new(attributes)
           compare_member(json_response, member)
         end
@@ -76,7 +73,7 @@ RSpec.describe 'Members', type: :request do
           create_member(attributes, @token)
         end
 
-        it 'should return a HTTP STATUS 400' do
+        it 'does return a HTTP STATUS 400' do
           expect(response).to have_http_status(:bad_request)
         end
 
@@ -86,11 +83,11 @@ RSpec.describe 'Members', type: :request do
     context "when user's not admin" do
       before { create_member(attributes, 'client_token') }
 
-      it 'should return a HTTP STATUS 401' do
+      it 'does return a HTTP STATUS 401' do
         expect(response).to have_http_status(:unauthorized)
       end
 
-      it 'should return a message error' do
+      it 'does return a message error' do
         expect(json_response[:message]).to eq('Unauthorized access.')
       end
     end
@@ -102,17 +99,16 @@ RSpec.describe 'Members', type: :request do
     context "when user's not admin" do
       before { update_member(@member.id, attributes, 'client_token') }
 
-      it 'should return a HTTP STATUS 401' do
+      it 'does return a HTTP STATUS 401' do
         expect(response).to have_http_status(:unauthorized)
       end
 
-      it 'should return a message error' do
+      it 'does return a message error' do
         expect(json_response[:message]).to eq('Unauthorized access.')
       end
     end
 
     context "when user's admin" do
-      
       before do
         admin_user = create(:admin_user)
         login_with_api(admin_user)
@@ -122,11 +118,12 @@ RSpec.describe 'Members', type: :request do
 
       context 'when params are valid' do
         before { update_member(@member.id, attributes, @token) }
-        it 'should return a HTTP STATUS 200' do
+
+        it 'does return a HTTP STATUS 200' do
           expect(response).to have_http_status(:ok)
         end
 
-        it "should return the updated member" do
+        it 'does return the updated member' do
           updated_member = Member.new(attributes)
           compare_member(json_response, updated_member)
         end
@@ -134,12 +131,13 @@ RSpec.describe 'Members', type: :request do
 
       context 'when params are empty' do
         before { update_member(@member.id, '', @token) }
-        it 'should return a HTTP STATUS 400' do
+
+        it 'does return a HTTP STATUS 400' do
           expect(response).to have_http_status(:bad_request)
         end
 
-        it 'should return an error message' do
-          expect(json_response[:error]).to eq("Parameter is missing or its value is empty")
+        it 'does return an error message' do
+          expect(json_response[:error]).to eq('Parameter is missing or its value is empty')
         end
       end
     end
@@ -151,16 +149,16 @@ RSpec.describe 'Members', type: :request do
     context "when user's not admin" do
       before { delete_member(@member.id, 'client_token') }
 
-      it 'should return a HTTP STATUS 401' do
+      it 'does return a HTTP STATUS 401' do
         expect(response).to have_http_status(:unauthorized)
       end
 
-      it 'should return an error message' do
-        expect(json_response[:message]).to eq("Unauthorized access.")
+      it 'does return an error message' do
+        expect(json_response[:message]).to eq('Unauthorized access.')
       end
     end
 
-    context "when user is admin" do
+    context 'when user is admin' do
       before do
         admin_user = create(:admin_user)
         login_with_api(admin_user)
@@ -170,36 +168,33 @@ RSpec.describe 'Members', type: :request do
 
       context 'when member is deleted' do
         before do |example|
-          unless example.metadata[:skip_before]
-            delete_member(@member.id, @token)
-          end
+          delete_member(@member.id, @token) unless example.metadata[:skip_before]
         end
 
-        it 'should delete member from database', :skip_before do
+        it 'does delete member from database', :skip_before do
           expect{ delete_member(@member.id, @token) }.to change(Member, :count).by(-1)
         end
 
-        it 'should return HTTP STATUS 200' do
+        it 'does return HTTP STATUS 200' do
           expect(response).to have_http_status(:ok)
         end
 
-        it 'should return a success message' do
-          expect(json_response[:message]).to eq("Succesfully deleted")
+        it 'does return a success message' do
+          expect(json_response[:message]).to eq('Succesfully deleted')
         end
       end
-      
+
       context "when member's not found" do
         before { delete_member(99, @token) }
 
-        it 'should return a HTTP STATUS 404' do
+        it 'does return a HTTP STATUS 404' do
           expect(response).to have_http_status(:not_found)
         end
 
-        it 'should return a member not found message' do
+        it 'does return a member not found message' do
           expect(json_response[:error]).to eq('member not found')
         end
       end
     end
   end
-  
 end
